fermctl/requirements.txt
Flask>=3.0
RPi.GPIO>=0.7; platform_machine == "aarch64" or platform_machine == "armv7l"
